% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_by_functions.R
\name{find_boxset}
\alias{find_boxset}
\title{Search by boxset title}
\usage{
find_boxset(boxset_title)
}
\arguments{
\item{boxset_title}{title of boxset}
}
\description{
Output a data frame for specific boxset
}
\examples{
library(criterioncollection)
find_boxset("The Before Trilogy")

}
