% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_by_functions.R
\name{find_by_director}
\alias{find_by_director}
\title{Search by movie director}
\usage{
find_by_director(film_director)
}
\arguments{
\item{film_director}{}
}
\description{
Output a data frame for specific movie director
}
\examples{
library(criterioncollection)
find_by_director("John Woo")

}
