% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_by_functions.R
\name{find_by_spine}
\alias{find_by_spine}
\title{Search by Spine Number}
\usage{
find_by_spine(spine_no)
}
\arguments{
\item{spine_no}{spine number}
}
\description{
Output a data frame for specific spine number
}
\examples{
library(criterioncollection)
find_by_spine("1")

}
