% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_by_functions.R
\name{find_by_title}
\alias{find_by_title}
\title{Search by film title}
\usage{
find_by_title(film_title)
}
\arguments{
\item{film_title}{}
}
\description{
Output a data frame for specific film title
}
\examples{
library(criterioncollection)
find_by_title("Amarcord")

}
