% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{criterion}
\alias{criterion}
\title{Data for movies in the Criterion Collection as of March 2022}
\format{
A data frame of 1,515 rows with each row representing an item, such
as a movie or box set, in the Criterion Collection.
\describe{
\item{spine}{Spine number of the movie or box set}
\item{year}{Year in which the movie was released}
\item{country}{Country in which the movie was made}
\item{title}{Movie or box set title}
\item{director}{Director of the movie}
}
}
\usage{
criterion
}
\description{
1,515 items in the Criterion Collection, including box sets, from
\url{https://www.criterion.com/shop/browse/list}.
}
\keyword{datasets}
